query_service/requirements.txt
Django==4.1.7
pymongo==3.12 ## Super importante este para evitar un erro que genera los modelos de Djongo al llamar get, create, etc
djangorestframework==3.14.0
opencv-python-headless==4.7.0.72
Pillow==9.5.0
pandas==2.0.1
requests==2.31.0
